[Fix] schedule 분리, 수정 요청 적용
